Test andando User andando

+++ VirtualTest/Services/Interfaces/IStrategyScore.cs
@@ -4,6 +4,6 @@ namespace VirtualTest.Services.Interfaces
 {
     interface IStrategyScore
     {
-        double GetScoreByOpenTDB(int amountCorrectAnswers, int amountQuestions,  Difficulty difficulty, int timeConsumed);
+        double GetScoreByOpenTDB(int amountCorrectAnswers, int amountQuestions,  Difficulty difficulty, double timeConsumed);
     }
 }
